java 5 does unboxing of value
git-svn-id: https://scm.mmbase.org/mmbase/trunk@28976 91430c5f-d668-0410-9e7d-dd663366dcfd

+++ contributions/CMSContainer/cmsc/sitemanagement/src/java/com/finalist/cmsc/beans/om/Portlet.java
@@ -1,6 +1,7 @@
 package com.finalist.cmsc.beans.om;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 
 import net.sf.mmapps.commons.beans.NodeBean;
 
@@ -43,7 +44,7 @@ public class Portlet extends NodeBean implements Comparable<Portlet> {
       if (views.isEmpty()) {
          return -1;
       }
-      return views.get(0).intValue();
+      return views.get(0);
    }
